Milestone: Complete Enterprise RBAC System

+++ app/Http/Controllers/ProductController.php
@@ -11,21 +11,28 @@ class ProductController extends Controller
     
      public function index(Request $request)
 {
-    //dd($request->all()); <-- mao ni ang gamiton unsaon pag test ang CRUD
+    if (! auth()->user()->hasPermission('products.view')) {
+        abort(403);
+    }
+
     if ($request->filled('search')) {
-    $search = $request->search;
-    $products = Product::where('name', 'LIKE', '%' . $search . '%')->get();
-} else {
-    $products = Product::with('category')->get();
-}
+        $search = $request->search;
+        $products = Product::where('name', 'LIKE', '%' . $search . '%')->get();
+    } else {
+        $products = Product::with('category')->get();
+    }
+
     $categories = Category::all();
+
     return view('products.index', compact('products', 'categories'));
 }
  
 
     public function store(ProductRequest $request)
 {
-  
+  if (! auth()->user()->hasPermission('products.create')) {
+    abort(403);
+}
 
     Product::create($request->validated());
 
@@ -34,12 +41,18 @@ class ProductController extends Controller
 
    public function edit(Product $product)
 {
+    if (! auth()->user()->hasPermission('products.edit')) {
+    abort(403);
+}
     $categories = Category::all();
     return view('products.edit', compact('product', 'categories'));
 }
 
     public function update(ProductRequest $request, Product $product)
 {
+    if (! auth()->user()->hasPermission('products.edit')) {
+    abort(403);
+}
 
     $product->update($request->validated());
 
@@ -50,6 +63,9 @@ class ProductController extends Controller
     public function destroy(Product $product)
 {
     
+if (! auth()->user()->hasPermission('products.delete')) {
+    abort(403);
+}
     $product->delete();
 
     return redirect('/products');
